requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
btcpay-python==1.3.0
certifi==2023.7.22
charset-normalizer==3.3.0
discord==2.3.2
discord.py==2.3.2
ecdsa==0.18.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.6
yarl==1.9.2
